[Refactor pt 2] - Create EditPanel, MenuPanel, and FuturePanel class
