+++ src/RestaurantList/GUI/listeners/ComboBoxListener.java
@@ -1,0 +1,14 @@
+package edu.ucsb.cs56.projects.utilities.restaurant_list;
+import javax.swing.*;
+import java.awt.event.*;
+
+/**
+ *  Get cuisine chosen by user and adds restaurant names
+ *  to the second JComboBox
+ */
+class ComboBoxListener implements ActionListener {
+	public void actionPerformed(ActionEvent event) {
+    	JComboBox cb = (JComboBox)event.getSource();
+	}
+}
+
